# Conflicts:
#	Desarrollo/Motor/DebugStatic/Motor.exp
#	Desarrollo/Motor/DebugStatic/Motor.lib
#	Desarrollo/Motor/Motor/DebugStatic/AnimationMesh.obj
#	Desarrollo/Motor/Motor/DebugStatic/AnimationReader.obj
#	Desarrollo/Motor/Motor/DebugStatic/Clock.obj
#	Desarrollo/Motor/Motor/DebugStatic/ClockImpl.obj
#	Desarrollo/Motor/Motor/DebugStatic/EngineDevice.obj
#	Desarrollo/Motor/Motor/DebugStatic/Input.obj
#	Desarrollo/Motor/Motor/DebugStatic/Motor.log
#	Desarrollo/Motor/Motor/DebugStatic/Motor.tlog/CL.command.1.tlog
#	Desarrollo/Motor/Motor/DebugStatic/Motor.tlog/CL.write.1.tlog
#	Desarrollo/Motor/Motor/DebugStatic/Motor.tlog/Motor.write.1u.tlog
#	Desarrollo/Motor/Motor/DebugStatic/Motor.tlog/link.command.1.tlog
#	Desarrollo/Motor/Motor/DebugStatic/Motor.tlog/link.read.1.tlog
#	Desarrollo/Motor/Motor/DebugStatic/Motor.tlog/link.write.1.tlog
#	Desarrollo/Motor/Motor/DebugStatic/ResourceManager.obj
#	Desarrollo/Motor/Motor/DebugStatic/SceneManager.obj
#	Desarrollo/Motor/Motor/DebugStatic/Shader.obj
#	Desarrollo/Motor/Motor/DebugStatic/TAnimation.obj
#	Desarrollo/Motor/Motor/DebugStatic/TAnimationGroupMesh.obj
#	Desarrollo/Motor/Motor/DebugStatic/TCamera.obj
#	Desarrollo/Motor/Motor/DebugStatic/TEntity.obj
#	Desarrollo/Motor/Motor/DebugStatic/TFlashLight.obj
#	Desarrollo/Motor/Motor/DebugStatic/TLuz.obj
#	Desarrollo/Motor/Motor/DebugStatic/TMesh.obj
#	Desarrollo/Motor/Motor/DebugStatic/TMeshGroup.obj
#	Desarrollo/Motor/Motor/DebugStatic/TModel.obj
#	Desarrollo/Motor/Motor/DebugStatic/TNode.obj
#	Desarrollo/Motor/Motor/DebugStatic/TPointLight.obj
#	Desarrollo/Motor/Motor/DebugStatic/TSunLight.obj
#	Desarrollo/Motor/Motor/DebugStatic/TTransform.obj
#	Desarrollo/Motor/Motor/DebugStatic/Texture.obj
#	Desarrollo/Motor/Motor/DebugStatic/Time.obj
#	Desarrollo/Motor/Motor/DebugStatic/main.obj

+++ Desarrollo/BulletTest/BulletTest/GUI/MenuGUI.cpp
@@ -228,7 +228,7 @@ bool MenuGUI::onAtrasClicked(const CEGUI::EventArgs & e)
 void MenuGUI::updateFondo(int velocidad)
 {
 	CEGUI::UDim suma;
-	suma.d_offset = velocidad;
+	suma.d_offset = (float) velocidad;
 
 	CEGUI::UVector2 newposition(CEGUI::UDim(imagen->getPosition().d_x) + suma, CEGUI::UDim(imagen->getPosition().d_y));
 	CEGUI::UVector2 newposition2(CEGUI::UDim(imagen2->getPosition().d_x) + suma, CEGUI::UDim(imagen2->getPosition().d_y));
